refactor(postgres-napi): keep native testing types separate
